+++ src/App.vue
@@ -5,29 +5,16 @@
   </div>
 </template>
 <script>
-  import Nav from '@/components/Nav'
-  import index from '@/components/index'
-  import chatrecord from '@/components/chatrecord'
-  import contact from '@/components/contact'
-  import userPu from '@/components/userPu'
-  export default {
-    name:'App',
-    data() {
-      return{
+import Nav from '@/components/Nav'
 
-      }
-    },
-    components:{
-     index, Nav,chatrecord,contact,userPu
-    },
-    methods:{
-
-    }
-  }
-
-
+export default {
+  name: 'App',
+  data() {
+    return {}
+  },
+  components: {
+    Nav
+  },
+  methods: {}
+}
 </script>
-
-<style>
-@import "~@/assets/fonts/iconfont.css";
-</style>
