+++ src/router/index.js
@@ -1,48 +1,18 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import index from '@/components/index'
-import chatrecord from '@/components/chatrecord'
-import contact from '@/components/contact'
-import userPu from '@/components/userPu'
+import routes from './router'
 
 Vue.use(VueRouter)
 
-const routes = [
-  {
-    path: '/',
-    component: index
-  },
-  {
-    path: '/index',
-    component: index
-  },
-  {
-    path: '/chatrecord',
-    component: chatrecord
-},
-  {
-    path: '/contact',
-    component: contact
-
-  },
-  {
-    path: '/userPu',
-    component: userPu
-  }
-
-    //path: '/about',
-    //name: 'About',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    //component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-
-]
-
 const router = new VueRouter({
-   mode: 'hash',
+  mode: 'history',
   base: process.env.BASE_URL,
   routes
 })
 
+router.beforeEach((to, from, next) => {
+  console.log(to);
+  console.log(from);
+  next();
+});
 export default router
